Cleanup and fix some credo alerts
