+++ host_core/lib/host_core/cloud_event.ex
@@ -1,4 +1,6 @@
 defmodule HostCore.CloudEvent do
+  @moduledoc false
+
   def new(data, event_type, host \\ HostCore.Host.host_key()) do
     stamp = DateTime.utc_now() |> DateTime.to_iso8601()
