+++ host_core/lib/host_core/control_interface.ex
@@ -1,2 +1,3 @@
 defmodule HostCore.ControlInterface do
+  @moduledoc false
 end
