+++ host_core/lib/host_core/lattice.ex
@@ -1,2 +1,3 @@
 defmodule HostCore.Lattice do
+  @moduledoc false
 end
