+++ host_core/lib/host_core/oci.ex
@@ -1,4 +1,6 @@
 defmodule HostCore.Oci do
+  @moduledoc false
+
   def allow_latest() do
     case :ets.lookup(:config_table, :config) do
       [config: config_map] -> config_map[:allow_latest]
