+++ host_core/lib/host_core/providers/builtin/logging.ex
@@ -1,4 +1,5 @@
 defmodule HostCore.Providers.Builtin.Logging do
+  @moduledoc false
   require Logger
 
   def invoke(actor, method, payload) when method in ["Logging.WriteLog", "WriteLog"] do
