+++ host_core/lib/host_core/wasmcloud.ex
@@ -1,2 +1,3 @@
 defmodule HostCore.WasmCloud do
+  @moduledoc false
 end
