+++ host_core/lib/host_core/web_assembly.ex
@@ -1,2 +1,3 @@
 defmodule HostCore.WebAssembly do
+  @moduledoc false
 end
