+++ host_core/test/e2e/kvcounter_test.exs
@@ -228,7 +228,6 @@ defmodule HostCore.E2E.KVCounterTest do
 
     HTTPoison.start()
     {:ok, resp} = HTTPoison.get("http://localhost:8081/foobar")
-    IO.inspect(resp)
 
     assert resp.body == "Guest call failed: Host error: Invocation not authorized\n"
     assert resp.status_code == 500
